+++ src/com/burksnet/code/games/rain/entity/Entity.java
@@ -6,10 +6,8 @@ import com.burksnet.code.games.rain.graphics.Screen;
 import com.burksnet.code.games.rain.level.Level;
 
 public abstract class Entity {
-
 	
-	
-	public int x, y;
+	public double x, y;
 	private boolean removed = false;
 	protected Level level;
 	protected final Random random = new Random();
